pyproject.toml
[build-system]
requires = ["hatchling>=1.21"]
build-backend = "hatchling.build"

[project]
name = "sessionhub-api"
version = "0.1.0"
description = "FastAPI + Strawberry GraphQL (Lambda via Mangum), Postgres on EC2, API Gateway HTTP+WS"
readme = "README.md"
requires-python = ">=3.12"
authors = [{ name = "You" }]
license = { text = "MIT" }

# Runtime deps (compatible, not overly pinned; Docker will lock later)
dependencies = [
  "fastapi>=0.115,<1",
  "strawberry-graphql[asgi]>=0.230,<1",
  "mangum>=0.17,<1",
  "uvicorn[standard]>=0.30,<1",
  "pydantic>=2.7,<3",
  "pydantic-settings>=2.4,<3",
  "structlog>=24.1,<25",
  "python-json-logger>=2.0,<3",
  # Drivers / future DB bits
  "asyncpg>=0.29,<1"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3,<9",
  "pytest-asyncio>=0.23,<1",
  "httpx>=0.27,<1",
  "mypy>=1.10,<2",
  "ruff>=0.5,<1",
  "types-python-dateutil"
]

[tool.hatch.build.targets.wheel]
packages = ["api"]

[tool.ruff]
line-length = 100
target-version = "py312"
select = ["E","F","I","UP","B","SIM","W","RUF"]
ignore = ["E501"]  # we'll manage line length pragmatically

[tool.mypy]
python_version = "3.12"
plugins = []
strict = false
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
no_implicit_optional = true
disallow_untyped_defs = true
check_untyped_defs = true
pretty = true